% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_utilities.R
\name{calculate_mode}
\alias{calculate_mode}
\title{Calculate Mode}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{A variable.}
}
\description{
Calculate mode of a caracter variable present in a data set.
}
\details{
\itemize{
\item Data must have a categorical variable you want to know the most frequent value.
}
}
\examples{
library(tidyverse)

 iris \%>\%
 mutate(cat_Sepal.Length=round(Sepal.Length, digits = 0)) \%>\%
 group_by(cat_Sepal.Length) \%>\%
 summarise(mode_species=calculate_mode(Species))

}
