% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_ML.R
\name{calibrate_and_plot}
\alias{calibrate_and_plot}
\title{ML Calibrate and Plot}
\usage{
calibrate_and_plot(..., type = "testing")
}
\arguments{
\item{...}{A fitted model object that is either:
\enumerate{
\item A modeltime table that has been created using \code{\link[=modeltime_table]{modeltime_table()}}
\item A workflow that has been fit by \href{https://workflows.tidymodels.org/reference/fit-workflow.html}{fit.workflow()} or
\item A parsnip model that has been fit using \href{https://parsnip.tidymodels.org/reference/fit.html}{fit.model_spec()}
}}

\item{type}{A test data set \code{tibble} containing future information (timestamps and actual values).}
}
\description{
Perform \href{https://cran.r-project.org/web/packages/modeltime/vignettes/getting-started-with-modeltime.html}{modeltime} calibrate and plot function at the same time.
}
\details{
-The results of calibration and plot are used to evaluate models
}
\examples{
library(tidyverse)
library(lubridate)
library(timetk)
library(parsnip)
library(rsample)
library(modeltime)

# Data
data_prepared_tbl <- m4_monthly \%>\% filter(id == "M750")

# Split Data 80/20
splits <- initial_time_split(data_prepared_tbl, prop = 0.9)

# Model: auto_arima
model_fit_arima <- arima_reg() \%>\%
    set_engine(engine = "auto_arima") \%>\%
    fit(value ~ date, data = training(splits))

# Calibrate and plot
calibrate_and_plot(model_fit_arima, type="testing")
}
