% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_utilities.R
\name{knit_with_date_html}
\alias{knit_with_date_html}
\title{Custom Knit function for RStudio}
\usage{
knit_with_date_html(input, ...)
}
\arguments{
\item{input}{the input file path}

\item{...}{other arguments passed to rmarkdown::render()}
}
\description{
Use to add date in the title of the knit markdown document
}
\examples{
#---
#knit: jofou.lib::knit_with_date_html
#---

}
