% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_EDA.R
\name{my_cat_dist}
\alias{my_cat_dist}
\title{EDA Categorical Variables Distribution}
\usage{
my_cat_dist(df, na.rm = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{na.rm}{will remove NAs before making plots.}
}
\description{
Show distribution of all categorical variable in data.
}
\details{
\itemize{
\item Data must have the type of variables your looking to summarize.
}
}
\examples{
library(tidyverse)

iris \%>\%
    my_cat_dist()

}
