% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_EDA.R
\name{my_corr_num_graph}
\alias{my_corr_num_graph}
\title{EDA Correlation Chart for all Numerics Variables}
\usage{
my_corr_num_graph(df, na.rm = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{na.rm}{will remove NAs before making plots.}
}
\description{
Show \href{https://github.com/braverock/PerformanceAnalytics}{PerformanceAnalytics} correlation chart for all the numeric varibles in the data.
}
\details{
\itemize{
\item Data must have the type of variables your looking to summarize.
}
}
\examples{
library(tidyverse)

iris \%>\%
    my_corr_num_graph()

}
