% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_EDA_inspect.R
\name{my_inspect_types}
\alias{my_inspect_types}
\title{EDA Inspect Data}
\usage{
my_inspect_types(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Use tu supply group names for the inspectdf function.}
}
\description{
Perform inspectdf function and put the results in preformated tables
}
\details{
\itemize{
\item Data must have the type fof variables your looking to summarize.
}
}
\examples{
library(tidyverse)

iris \%>\%
    my_inspect_types()

}
