% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_EDA.R
\name{my_num_dist}
\alias{my_num_dist}
\title{EDA Numerical Variables Distribution}
\usage{
my_num_dist(df, na.rm = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{na.rm}{will remove NAs before making plots.}
}
\description{
Show distribution of numeric variables with a supperposition of box_plot and an histogram.
}
\details{
\itemize{
\item Data must have the type of variables your looking to summarize.
}
}
\examples{
library(tidyverse)

iris \%>\%
    my_num_dist()

}
