% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jofou_utilities.R
\name{rle_fx}
\alias{rle_fx}
\title{Compter le nombre de répétition consécutives}
\usage{
rle_fx(x)
}
\arguments{
\item{x}{column name}
}
\description{
Use to count the number of repetition of same string in a dataframe column
}
